requirements.txt
about-time==4.2.1
alive-progress==3.1.4
colorama==0.4.6
Faker==19.12.1
grapheme==0.6.0
psycopg==3.1.12
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.8.0
